refactor: move class & Ladder 에서 Line 에 대한 의존성 제거
