+++ src/test/java/nextstep/ladder/domain/LadderTest.java
@@ -4,9 +4,11 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.util.List;
 import java.util.stream.Stream;
 
 import static java.util.Arrays.asList;
+import static java.util.stream.Collectors.toList;
 import static nextstep.ladder.domain.IndexedName.wrap;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.entry;
@@ -20,8 +22,12 @@ class LadderTest {
 
     @BeforeEach
     void setUp() {
-        line = Ladder.of(Stream.of(Spoke.of(true, false), Spoke.of(false, true)),
+        line = of(Stream.of(Spoke.of(true, false), Spoke.of(false, true)),
                          wrap(asList("boom", "5000", "boom")));
+    }
+    public static Ladder of(Stream<Spoke> spokes, List<IndexedName> goals) {
+        return new Ladder(new DefaultLadderLine(spokes.map(Line::new)
+                                                               .collect(toList())), goals);
     }
 
     @DisplayName("특정 플레이어의 이동결과를 리턴한다")
